// QuickApp.Pro/ClientApp/app/models/assettypesinglescreen.model.ts
export class AssetTypeSingleScreen {
    assetTypeSingleScreenId: number;
    assetTypeSingleId: string;
    assetTypeName: string;
    assetTypeMemo: string;
    createdBy: string;
    updatedBy: string;
    createdDate: Date;
    updatedDate: Date;
    isDelete: boolean;
    isActive: boolean;
    assetTypeList: AssetTypeSingleScreen[];

}